+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/ui-test-utils
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [[`dc63ac61ce5768fdbc353800f4477f3fc71c1970`](https://github.com/aws-amplify/amplify-ui/commit/dc63ac61ce5768fdbc353800f4477f3fc71c1970)]:
+  - @aws-amplify/ui-react-storage@3.13.0
+
 ## 0.0.9
 
 ### Patch Changes